Ej herencia

+++ src/UT4/herencia/ejHerencias2/VehiculosConMotor.java
@@ -3,9 +3,9 @@ package UT4.herencia.ejHerencias2;
 public class VehiculosConMotor extends Vehiculos{
 
     private int cilindrada;
-    private int combustible;
+    private double combustible;
 
-    public VehiculosConMotor(String color, String modelo, double precio, String marca, int cilindrada, int combustible) {
+    public VehiculosConMotor(String color, String modelo, double precio, String marca, int cilindrada, double combustible) {
         super(color, modelo, precio, marca);
         this.cilindrada = cilindrada;
         this.combustible = combustible;
